updated cat6egory Controller

+++ app/Http/Controllers/BlogpostCategoryController.php
@@ -33,26 +33,44 @@ class BlogpostCategoryController extends Controller
     public function viewCategories()
     {
         $data = Category::paginate(3);
-        return $data;
+        if ($data) {
+            return $this->sendNotSuccessResponse(__('Success'), $data);
+        } else {
+            return $this->sendFailedResponse(__('Success'));
+        }
     }
 
     public function showCategoryDetails($id)
     {
         $data = Category::where($id, 'id')->first();
-        return $data;
+        if ($data) {
+            return $this->sendNotSuccessResponse(__('Success'), $data);
+        } else {
+            return $this->sendFailedResponse(__('Success'));
+        }
     }
 
     public function updateCategory($id, $request)
     {
         $category = Category::findOrFail($id);
-        $category->update([
-            'category_name' => $request->input('category_name'),
-        ]);
+        if ($category) {
+            $category->update([
+                'category_name' => $request->input('category_name'),
+            ]);
+            return $this->sendNotSuccessResponse(__('Requested Category Updated Successfully'));
+        } else {
+            return $this->sendNotFoundResponse(__('Requested Category Not Found'));
+        }
     }
 
     public function deleteCategory($id)
     {
         $category = Category::findOrFail($id);
-        $category->delete();
+        if ($category) {
+            $category->delete();
+            return $this->sendSuccessResponse(__('Requested Category Deleted Successfully'));
+        } else {
+            return $this->sendNotFoundResponse(__('Requested Category Not Found'));
+        }
     }
 }
